third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-item-start-before-get-set-expected.txt

Test that setting and getting grid-column-start and grid-row-start works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-column-start and grid-row-start set through CSS
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-row') is 'auto / auto'
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / auto'

Test getting and setting grid-column-start and grid-row-start through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue

Test setting grid-column-start and grid-row-start to 'inherit' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue

Test setting grid-column-start and grid-row-start to 'initial' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue

Test setting grid-column-start and grid-row-start back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is '18'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is '18 / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is '66'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is '66 / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / auto'
PASS successfullyParsed is true

TEST COMPLETE
